+++ prisma/migrations/20231012060215_multiple_webhook_types/migration.sql
@@ -1,0 +1,9 @@
+/*
+  Warnings:
+
+  - The `type` column on the `Webhook` table would be dropped and recreated. This will lead to data loss if there is data in the column.
+
+*/
+-- AlterTable
+ALTER TABLE "Webhook" DROP COLUMN "type",
+ADD COLUMN     "type" TEXT[];