+++ src/auth/auth.controller.ts
@@ -123,7 +123,7 @@ export class AuthController {
   }
 
   @ApiOperation({ summary: 'Sends link to resets password of the user' })
-  @Get('reset-password')
+  @Post('request-reset-password')
   async resetPasswordEmail(
     @Body() resetPasswordRequest: ResetPasswordEmailRequest,
   ): Promise<void> {