+++ src/webhook/test/webhook.factory.ts
@@ -6,7 +6,7 @@ const MAX_I32 = 2147483647;
 type CreateWebhookRequestOverrides = {
   id?: number;
   name?: string;
-  type?: WebhookEventType;
+  types?: WebhookEventType[];
   url?: string;
   token?: string;
 };
@@ -15,7 +15,7 @@ export const createWebhookRequest = (
   overrides: CreateWebhookRequestOverrides = {},
 ) => ({
   name: overrides.name ?? faker.word.noun(),
-  type: overrides.type ?? WebhookEventType.RecipeCreated,
+  types: overrides.types ?? [WebhookEventType.RecipeCreated],
   url: overrides.url ?? faker.internet.url(),
   token: overrides.token ?? faker.string.alphanumeric(32),
 });
@@ -26,7 +26,7 @@ export const createWebhookWithUserId = (
 ) => ({
   id: overrides.id ?? faker.number.int({ max: MAX_I32 }),
   name: overrides.name ?? faker.word.noun(),
-  type: overrides.type ?? WebhookEventType.RecipeCreated,
+  types: overrides.types ?? [WebhookEventType.RecipeCreated],
   url: overrides.url ?? faker.internet.url(),
   token: overrides.token ?? faker.string.alphanumeric(32),
   userId,
@@ -39,7 +39,7 @@ export function createWebhookResponse() {
     id: faker.number.int({ max: MAX_I32 }),
     name: faker.word.noun(),
     url: faker.internet.url(),
-    type: WebhookEventType.RecipeCreated,
+    types: [WebhookEventType.RecipeCreated],
     token: '',
     initVector: '',
     authTag: '',