+++ src/webhook/webhook.repository.ts
@@ -7,7 +7,7 @@ import { CreateWebhookRequest, WebhookEventType } from './dto';
 export class WebhookRepository {
   constructor(private prisma: PrismaService) {}
 
-  async createWebhook(
+  createWebhook(
     userId: number,
     webhookData: CreateWebhookRequest,
     iv?: string,
@@ -23,7 +23,7 @@ export class WebhookRepository {
     });
   }
 
-  async createWebhookEvent(
+  createWebhookEvent(
     webhookId: number,
     data: any,
     webhookType: WebhookEventType,
@@ -37,7 +37,7 @@ export class WebhookRepository {
     });
   }
 
-  async getDataToSendWebhookEvent(webhookId: number) {
+  getDataToSendWebhookEvent(webhookId: number) {
     return this.prisma.webhook.findUnique({
       where: {
         id: webhookId,
@@ -51,7 +51,7 @@ export class WebhookRepository {
     });
   }
 
-  async getAllWebhooksByUserId(userId: number): Promise<Webhook[]> {
+  getAllWebhooksByUserId(userId: number): Promise<Webhook[]> {
     return this.prisma.webhook.findMany({
       where: {
         userId,
@@ -59,7 +59,7 @@ export class WebhookRepository {
     });
   }
 
-  async getWebhookById(webhookId: number): Promise<Webhook> {
+  getWebhookById(webhookId: number): Promise<Webhook> {
     return this.prisma.webhook.findUnique({
       where: {
         id: webhookId,